cub3d to norminette and change makefile

+++ cub3d/libft/ft_substr.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char		*sub;
-	size_t		i;
+	int     	i;
 
 	if (!s)
 		return (0);
